Refactor `MultipleTEnumValues` cop to use `TEnum` mixin
