+++ lib/rubocop/cop/sorbet_cops.rb
@@ -20,7 +20,6 @@ require_relative "sorbet/redundant_extend_t_sig"
 require_relative "sorbet/type_alias_name"
 require_relative "sorbet/obsolete_strict_memoization"
 require_relative "sorbet/buggy_obsolete_strict_memoization"
-require_relative "sorbet/multiple_t_enum_values"
 
 require_relative "sorbet/rbi/forbid_extend_t_sig_helpers_in_shims"
 require_relative "sorbet/rbi/forbid_rbi_outside_of_allowed_paths"
@@ -43,5 +42,7 @@ require_relative "sorbet/sigils/strict_sigil"
 require_relative "sorbet/sigils/strong_sigil"
 require_relative "sorbet/sigils/enforce_sigil_order"
 require_relative "sorbet/sigils/enforce_single_sigil"
+
+require_relative "sorbet/t_enum/multiple_t_enum_values"
 
 require_relative "sorbet/mutable_constant_sorbet_aware_behaviour"
