+++ spec/rubocop/cop/sorbet/t_enum/multiple_t_enum_values_spec.rb
